requirements.txt
anyio==3.6.2
asgiref==3.4.1
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cryptography==39.0.1
dataclasses==0.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.83.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.13.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.0.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
starlette==0.19.1
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.14
uvicorn==0.16.0
zipp==3.6.0
